move test environment into fixture, fix injection issues in CI, justfile updates
