+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+# Default: everything
+* @bckohan
