+++ .github/workflows/lint.yml
@@ -69,8 +69,10 @@ jobs:
       - name: Install Just
         uses: extractions/setup-just@e33e0265a09d6d736e2ee1e0eb685ef1de4669ff
       - name: Install Dependencies
+        env:
+          PYTHON_PATH: ${{ steps.sp.outputs.python-path }}
         run: |
-          just setup ${{ steps.sp.outputs.python-path }}
+          just setup "$PYTHON_PATH"
           if [[ "${{ matrix.django-version }}" =~ (a|b|rc) ]]; then
             just test-lock Django==${{ matrix.django-version }}
           else
@@ -88,9 +90,9 @@ jobs:
         timeout-minutes: 60
       - name: Run Static Analysis
         run: |
-          just test ./tests/test_verify_environment.py
           just check-lint
           just check-format
           just check-types
           just check-package
           just check-readme
+          just check-docs
