+++ .github/workflows/release.yml
@@ -46,11 +46,11 @@ jobs:
       uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405
       id: sp
       with:
-        python-version: "==3.14"  # for tomlib
+        python-version: "3.14"  # for tomlib
     - name: Install uv
       uses: astral-sh/setup-uv@eac588ad8def6316056a12d4907a9d4d84ff7a3b
       with:
-        enable-cache: true
+        enable-cache: false
         restore-cache: false
         save-cache: false
     - name: Setup Just
