+++ .gitignore
@@ -138,3 +138,4 @@ type_check.py
 
 .DS_Store
 CLAUDE.md
+uv.lock
