+++ pyproject.toml
@@ -109,8 +109,7 @@ typing = [
     "django-stubs[compatible-mypy]>=5.1.3",
     "django-stubs-ext>=5.1.3",
     "types-django-filter>=25.2.0.20251228",
-    "djangorestframework-stubs>=3.16.1",
-    "types-pyyaml>=6.0.12.20241230",
+    "djangorestframework-stubs>=3.16.1"
 ]
 lint = [
     "ruff>=0.15.0",
